requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
requests==2.21.0
firebase-admin==2.14.0
PyMySQL==0.8.0
google-auth==1.6.2
sendgrid==5.3.0
